requirements.txt
antlr4-python3-runtime==4.11.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==2.2.2
packaging==25.0
pycparser==2.22
python-dotenv==1.1.0
SQLAlchemy==2.0.41
sympy==1.13.3
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
Pillow